Add a D2 manifest service
